backend/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
google-generativeai==0.8.3
python-dotenv==1.0.0
supabase==2.3.4
fastapi==0.116.1
uvicorn==0.35.0
aiohttp==3.9.1 